backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML Libraries
anthropic>=0.43.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4
torch==2.1.0
transformers==4.36.0
sentence-transformers==2.2.2

# Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring and Metrics
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
tzdata==2023.3
ujson==5.8.0
orjson==3.9.10

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0

# WebSocket
websockets==12.0
starlette==0.27.0

# HTTP Client
aiohttp==3.9.1

# CORS
fastapi-cors==0.1.0

# Compression
brotli==1.1.0

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# File Uploads
python-magic==0.4.27

# Email
aiosmtplib==2.0.0

# Rate Limiting
slowapi==0.1.8

# Health Checks
healthchecks==2.0.0